+++ NEWS.md
@@ -1,3 +1,9 @@
+# sjlabelled 1.0.5
+
+## General
+
+* Remove unnecessary imports.
+
 # sjlabelled 1.0.4
 
 ## Changes to functions